Déplace les règles dans `./regles`
